+++ events/clientReady.js
@@ -1,0 +1,10 @@
+const { Events } = require('discord.js');
+const logger = require('../utility/logger');
+
+module.exports = {
+	name: Events.ClientReady,
+	once: true,
+	execute(client) {
+		logger.info(`Ready! Logged in as ${client.user.tag}`);
+	},
+};
